Features/diary

+++ src/main/java/work/appdeploys/equipmentcontrolsystem/controllers/OrdersControllers.java
@@ -6,24 +6,17 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import work.appdeploys.equipmentcontrolsystem.constants.MessageResource;
-import work.appdeploys.equipmentcontrolsystem.models.dtos.OrdersRequestDto;
+import work.appdeploys.equipmentcontrolsystem.models.dtos.OrderResponseDto;
 import work.appdeploys.equipmentcontrolsystem.models.structures.ExcelDto;
 import work.appdeploys.equipmentcontrolsystem.models.structures.OrdersResponse;
+import work.appdeploys.equipmentcontrolsystem.services.FileCopy;
 import work.appdeploys.equipmentcontrolsystem.services.OrdersService;
 
 import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 import java.io.IOException;
-import java.nio.file.Files;
 import java.time.LocalDate;
 import java.util.Arrays;
 @Tag(name="ordes")
@@ -33,20 +26,21 @@ import java.util.Arrays;
 @Slf4j
 public class OrdersControllers {
     private final OrdersService ordersService;
+    private final FileCopy fileCopy;
 
     @PostMapping()
-    public ResponseEntity<OrdersResponse> save(@RequestBody @Valid OrdersRequestDto ordersRequestDto){
+    public ResponseEntity<OrdersResponse> save(@RequestBody @Valid OrderResponseDto orderResponseDto){
         try{
-            return ResponseEntity.ok(new OrdersResponse(MessageResource.ORDER_SAVED, Arrays.asList(ordersService.save(ordersRequestDto))));
+            return ResponseEntity.ok(new OrdersResponse(MessageResource.ORDER_SAVED, Arrays.asList(ordersService.save(orderResponseDto))));
         }catch (Exception ex){
             return ResponseEntity.badRequest().body(new OrdersResponse(ex.getMessage(), Arrays.asList()));
         }
     }
 
     @PutMapping
-    public ResponseEntity<OrdersResponse> update(@RequestBody @Valid OrdersRequestDto ordersRequestDto){
+    public ResponseEntity<OrdersResponse> update(@RequestBody @Valid OrderResponseDto orderResponseDto){
         try{
-            return ResponseEntity.ok(new OrdersResponse(MessageResource.ORDER_UPDATED,Arrays.asList(ordersService.update(ordersRequestDto))));
+            return ResponseEntity.ok(new OrdersResponse(MessageResource.ORDER_UPDATED,Arrays.asList(ordersService.update(orderResponseDto))));
         }catch (Exception ex){
             return ResponseEntity.badRequest().body(new OrdersResponse(ex.getMessage(), Arrays.asList()));
         }
@@ -91,19 +85,22 @@ public class OrdersControllers {
 
     @GetMapping(path = "/excelorders/{dateTo}/{idSchool}")
     public void ordersExceclRespor(HttpServletResponse response, @PathVariable("dateTo") @DateTimeFormat(pattern = "yyyy-MM-dd") LocalDate dateTo, @PathVariable("idSchool") Long idSchool) throws IOException {
-        ExcelDto excelDto = ordersService.excelOrders(dateTo, idSchool);
+        ExcelDto excelDto = null;
         try {
+            excelDto = ordersService.excelOrders(dateTo, idSchool);
             String headerVal = "attachment; filename=" + excelDto.getNameExcel();
             response.setHeader("Content-Disposition", headerVal);
             response.setContentType(MediaType.APPLICATION_OCTET_STREAM.getType());
-            Files.copy(excelDto.getTmpExcel().toPath(), response.getOutputStream());
+            fileCopy.copyFile(excelDto.getTmpExcel(), response.getOutputStream());
 
             response.setStatus(HttpServletResponse.SC_OK);
         } catch (Exception e) {
             response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
         }finally {
             try {
-                excelDto.getTmpExcel().delete();
+                if(excelDto != null){
+                    excelDto.getTmpExcel().delete();
+                }
             }catch (Exception e){
                 log.error("Error delete file temporal");
             }
